docs(storybook): upgrade form story to storybook 5
